Inline RPC binding generation in main
